Add Solidity, WallETH, KEthereum
replacement for #45 as the file changed

btw: I also think it could make sense to distinguish between "EIP not relevant" and "work not started" - currently they are handled the same - but we would get an better overview by seeing them as separate things IMHO
e.g. for solidity 1559 is not relevant - but 3198 is. Now this can only be represented by - | -