Strip docstrings; undoc parameters (closes #26)
Since the docstrings are rendered as HTML, there's really no reason to
include them in the code snippets. This commit adds a dependency on
undocliteral, a hacky tool to strip docstrings when using
literalinclude.

This commit also removes the types from parameters in docstrings. The
signatures are annotated and checked with mypy, so including the types
in the docstrings is just duplicated information that can get out of
sync.