Add tool that enforces patch hygiene between hardforks.
Ideally we want to support meaningful diffs between different hardforks,
and generating meaningful diffs is a difficult problem.

This tool scans the sources of all hardforks and ensures that the
relative order of functions, classes, and constants doesn't change
between forks.

Hopefully this'll give us enough similarity for `rstdiff` to work its
magic.