Fixes transaction parameters for different transaction types.
Note: The `type` should be a single value literal for each transaction type, but I don't know how to do that.  I moved the type into the specific transactions in hopes that someone will be able to fix it to be the appropriate literal instead of a byte for each.