Migrate away from assert keyword for non-bugs. (closes #320)
The assert keyword in python can be disabled with the `-O` flag. Since
many of the assertions in the spec are actually intended to be runtime
checks, using assert doesn't make sense.

This commit adds an `ensure` function, which pretty much works the same
way as assert, but doesn't get optimized out.