update test collection
This commit updates the test collection logic to collect the tests
for a particular fork by looking at the  attribute instead of the
test key. This collects some additional tests that werent being collected before.
Some of these tests, specifically the  currently fail but have been
added to the ignore list. This will need to be explored in depth in the future