fetch fixtures at pytest sessionstart
This change fetches the test fixtures from the relevant repositories/releases
as and when required. As pytest is run, it checks at the beginning of the session
if the test fixtures are already available. If not, it fetches them. The following forms
of fetching are supported

1. Download the fixtures from a certain relase artifact (.tar.gz)
2. Clone the fixtures given a repository link and commit. Submodules will also be recursively
cloned in this case.

The fetching process is made safe for parallel process runs (with xdist for example) by
using a signal file which essentially tells the other processes that a process is already
fetching the relevant fixtures and they just have to wait.

The test run phase starts only once all the fixtures have been downloaded/cloned. In order to change
the links/commit hashes of the fixtures update tests/conftest.py