refactor tests
Currently the testing framework creates the Block and Genesis Block datatypes  from the various
fields available in the fixture json files. There are a couple of problems with this approach

1. The ethereum tests repository looks upon a lot of these json fields as an aid for human
   readability and looks at the rlp fields as the primary source of the data.
2. If new fields are added to the Block, Header, Transaction and Withdrawal  datatypes,  patches
   will have to be added to fix the testing logic.

Hence, the specs need to derive the data structures for testing by decoding the rlp instead of
using the json fields. This will also serve as an additional layer of testing for the rlp decode
logic within the specs. The current commit makes that change and also re-factors some of the code
in the state test helper.