refactor base fee calculation
The base fee calculation should be separated into its own function so that the t8n tool can call that function in order to derive the base_fee_per_gas from the parent block parameters. This commit performs this refactor and also simplifies the calculation.

The commit also corrects a small bug in Arrow Glacier and Gray Glacier. Currently, the fork block in these forks has INITIAL_BASE_FEE as the base fee. However, this was originally meant to be true only in case of London