doc flow: fix automatic filler doc generation
The mkdocs-gen-files plugin wasn't being used correctly (it wrote to files directly within a tmp ./docs/ directory instead of writing them using `mkdocs_gen_files.open()`). This caused a couple of issues with the doc build flow that are now resolved. The required change for navigation content was incompatible with mkdocs-awesome-pages-plugin; now mkdocs-literate-nav is used to achieve the desired ordering in nav instead.