pytest: enable fork spec via command-line and fork validity markers
Previously, tests had to be explicitly parametrized by fork by the test implementer. Now the implementer specifies:
1. Optionally, the fork range to parametrize over on the command-line via --from=FORK1 and --until=FORK2 OR --fork=FORK.
2. Marks the test case/function/class/module as pytest.mark.valid_from(FORKX)/pytest.mark.valid_until(FORKY)/pytest.mark.valid_only_for(FORKZ).

Due to the special behaviour of pytest with markers that take arguments that are callable, the FORK is specified to these markers as a string (they are checked in pytest_generate_tests()) instead of specifying a fork class from ethereum_test_forks. This makes it easier for the implementer (less import clutter), imo. Regarding the pytest idiosyncracy, if the class itself would be provided as the marker argument, then pytest tries to call this value. There is a work-around, which is to write @pytest.mark.valid_from.with_args(FORK) which is preventively clunky, imo; see https://docs.pytest.org/en/latest/example/markers.html#passing-a-callable-to-custom-markers for more info.