update to latest test releases + fixes
This commit updates to the latest test releases and fixes some minor bugs exposed by the increased coverage.
In particular, the bugs related to faulty return flag in the type section are fixed. Any code section that has
RETF instruction or has a JUMPF to a returning section is considered a returning section. Otherwise, the code section
is considered non-returning. The output in the type section should reflect this behaviour. If not, the container is
invalid