fix(consume): fix loading of multi-index, mixed tx type fixtures (#1330)
* fix(consume): fix loading of multi-index mixed tx type fixtures

Test fixtures from ethereum/{tests,legacyTests} can be parametrized by data, gas and value index. In this case, the tx type can be either be an access list type transaction (`0x1`), or not (`0x0`). For the latter, the entry in the transaction `accessList` field is set to `null` (an empty list can also be specified, but then this is a `0x1` type tx). This change handles this case.

* docs: update changelog

* docs: update pr link in changelog