Improve fixture fetching
Hoists the file system locks to the session level (instead of only while
fetching) so the whole fixtures folder is protected while the tests run.
Previously, running a second instance of the tests could swap out the
fixtures while the first instance is using them.

Re-download and re-extract HTTP tarball fixtures for every pytest
session. Previously it was assumed that if the directory exists, it was
the correct version of the fixtures. Now the fixtures are guaranteed to
be up-to-date.

HTTP responses are now cached in the user's home directory to avoid
re-downloading from GitHub when possible. The cache gets cleaned during
each pytest session, so it isn't particularly effective when changing
versions, but at least runs with the same version don't waste bandwidth.